Add assert in Main class that jude cannot be null
There is an assumption that the default file path is correct. In
some cases where the parent directory is not present, the program
will output IOException when loading the save file.

Asserting the path existence assures the correct code behaviour.

As a step towards such correction, let's assert the jude instance
cannot be null.