extract EditCommand's execute() cases into  4 different classes
EditCommand's execute() method parses the type of edit to give editType (which should be done by the parser) and executes based on the editType, this results in a deeply nested code and a very long method.

Refactoring this command into 4 different command classes allows the parser to determine which type of edit command to generate, while shortening the execute() method.

Let's:
* Remove the EditCommand class
* Add 4 different classes EditDescriptionCommand for ("/description"), EditToBeCompletedByCommand for ("/by"), EditStartTimeCommand for ("/start)", EditEndTimeCommand for ("/end)", each with an execute() method containing the functions of the respective of the 4 cases of editType
* Edit the Parser's parse() method to create each of these commands given their respective cases
* Allow ToDoList's getTask() method to throw an IndexOutOfBoundsException so that we know if we failed to get a task and allow each of our new command Classes to handle the error by informing that the task does not exist